move into beacon server
