+++ beacon-chain/rpc/service.go
@@ -3,7 +3,6 @@ package rpc
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net"
 	"time"
@@ -125,7 +124,14 @@ func (s *Service) Start() {
 		s.grpcServer = grpc.NewServer()
 	}
 
-	pb.RegisterBeaconServiceServer(s.grpcServer, s)
+	beaconServer := &BeaconServer{
+		beaconDB: s.beaconDB,
+		ctx: s.ctx,
+		attestationService: s.attestationService,
+		incomingAttestation: s.incomingAttestation,
+		canonicalStateChan: s.canonicalStateChan,
+	}
+	pb.RegisterBeaconServiceServer(s.grpcServer, beaconServer)
 	pb.RegisterValidatorServiceServer(s.grpcServer, s)
 	pb.RegisterProposerServiceServer(s.grpcServer, s)
 	pb.RegisterAttesterServiceServer(s.grpcServer, s)
@@ -156,58 +162,6 @@ func (s *Service) Stop() error {
 // TODO(1205): Add service health checks.
 func (s *Service) Status() error {
 	return nil
-}
-
-// CanonicalHead of the current beacon chain. This method is requested on-demand
-// by a validator when it is their time to propose or attest.
-func (s *Service) CanonicalHead(ctx context.Context, req *ptypes.Empty) (*pbp2p.BeaconBlock, error) {
-	block, err := s.beaconDB.ChainHead()
-	if err != nil {
-		return nil, fmt.Errorf("could not get canonical head block: %v", err)
-	}
-	return block, nil
-}
-
-// CurrentAssignmentsAndGenesisTime returns the current validator assignments
-// based on the beacon node's current, canonical crystallized state.
-// Validator clients send this request once upon establishing a connection
-// to the beacon node in order to determine their role and assigned slot
-// initially. This method also returns the genesis timestamp
-// of the beacon node which will allow a validator client to setup a
-// a ticker to keep track of the current beacon slot.
-func (s *Service) CurrentAssignmentsAndGenesisTime(
-	ctx context.Context,
-	req *pb.ValidatorAssignmentRequest,
-) (*pb.CurrentAssignmentsResponse, error) {
-	beaconState, err := s.beaconDB.State()
-	if err != nil {
-		return nil, fmt.Errorf("could not get beacon state: %v", err)
-	}
-	var keys []*pb.PublicKey
-	if req.AllValidators {
-		for _, val := range beaconState.ValidatorRegistry {
-			keys = append(keys, &pb.PublicKey{PublicKey: val.Pubkey})
-		}
-	} else {
-		keys = req.PublicKeys
-		if len(keys) == 0 {
-			return nil, errors.New("no public keys specified in request")
-		}
-	}
-	assignments, err := assignmentsForPublicKeys(keys, beaconState)
-	if err != nil {
-		return nil, fmt.Errorf("could not get assignments for public keys: %v", err)
-	}
-
-	timestamp, err := ptypes.TimestampProto(time.Unix(int64(beaconState.GenesisTime), 0))
-	if err != nil {
-		return nil, fmt.Errorf("could not create timestamp proto object %v", err)
-	}
-
-	return &pb.CurrentAssignmentsResponse{
-		GenesisTimestamp: timestamp,
-		Assignments:      assignments,
-	}, nil
 }
 
 // ProposeBlock is called by a proposer in a sharding validator and a full beacon node
@@ -311,27 +265,6 @@ func (s *Service) AttestHead(ctx context.Context, req *pb.AttestRequest) (*pb.At
 	return &pb.AttestResponse{AttestationHash: h[:]}, nil
 }
 
-// LatestAttestation streams the latest processed attestations to the rpc clients.
-func (s *Service) LatestAttestation(req *ptypes.Empty, stream pb.BeaconService_LatestAttestationServer) error {
-	sub := s.attestationService.IncomingAttestationFeed().Subscribe(s.incomingAttestation)
-	defer sub.Unsubscribe()
-	for {
-		select {
-		case attestation := <-s.incomingAttestation:
-			log.Info("Sending attestation to RPC clients")
-			if err := stream.Send(attestation); err != nil {
-				return err
-			}
-		case <-sub.Err():
-			log.Debug("Subscriber closed, exiting goroutine")
-			return nil
-		case <-s.ctx.Done():
-			log.Debug("RPC context closed, exiting goroutine")
-			return nil
-		}
-	}
-}
-
 // ValidatorShardID is called by a validator to get the shard ID of where it's suppose
 // to proposer or attest.
 func (s *Service) ValidatorShardID(ctx context.Context, req *pb.PublicKey) (*pb.ShardIDResponse, error) {
@@ -393,100 +326,8 @@ func (s *Service) ValidatorIndex(ctx context.Context, req *pb.PublicKey) (*pb.In
 	return &pb.IndexResponse{Index: index}, nil
 }
 
-// ValidatorAssignments streams validator assignments every cycle transition
-// to clients that request to watch a subset of public keys in the
-// CrystallizedState's active validator set.
-func (s *Service) ValidatorAssignments(
-	req *pb.ValidatorAssignmentRequest,
-	stream pb.BeaconService_ValidatorAssignmentsServer) error {
-
-	sub := s.chainService.CanonicalStateFeed().Subscribe(s.canonicalStateChan)
-	defer sub.Unsubscribe()
-	for {
-		select {
-		case beaconState := <-s.canonicalStateChan:
-			log.Info("Sending new cycle assignments to validator clients")
-
-			var keys []*pb.PublicKey
-			if req.AllValidators {
-				for _, val := range beaconState.ValidatorRegistry {
-					keys = append(keys, &pb.PublicKey{PublicKey: val.Pubkey})
-				}
-			} else {
-				keys = req.PublicKeys
-				if len(keys) == 0 {
-					return errors.New("no public keys specified in request")
-				}
-			}
-
-			assignments, err := assignmentsForPublicKeys(keys, beaconState)
-			if err != nil {
-				return fmt.Errorf("could not get assignments for public keys: %v", err)
-			}
-
-			// We create a response consisting of all the assignments for each
-			// corresponding, valid public key in the request. We also include
-			// the beacon node's current beacon slot in the response.
-			res := &pb.ValidatorAssignmentResponse{
-				Assignments: assignments,
-			}
-			if err := stream.Send(res); err != nil {
-				return err
-			}
-		case <-sub.Err():
-			log.Debug("Subscriber closed, exiting goroutine")
-			return nil
-		case <-s.ctx.Done():
-			log.Debug("RPC context closed, exiting goroutine")
-			return nil
-		}
-	}
-}
-
 // ValidatorEpochAssignments ... WIP
 func (s *Service) ValidatorEpochAssignments(ctx context.Context, req *pb.ValidatorEpochAssignmentsRequest) (*pb.ValidatorEpochAssignmentsResponse, error) {
 	return nil, nil
 }
 
-// assignmentsForPublicKeys fetches the validator assignments for a subset of public keys
-// given a crystallized state.
-func assignmentsForPublicKeys(keys []*pb.PublicKey, beaconState *pbp2p.BeaconState) ([]*pb.Assignment, error) {
-	// Next, for each public key in the request, we build
-	// up an array of assignments.
-	assignments := []*pb.Assignment{}
-
-	for _, val := range keys {
-		if len(val.PublicKey) == 0 {
-			continue
-		}
-		// For the corresponding public key and current crystallized state,
-		// we determine the assigned slot for the validator and whether it
-		// should act as a proposer or attester.
-		assignedSlot, role, err := v.ValidatorSlotAndRole(
-			val.PublicKey,
-			beaconState.ValidatorRegistry,
-			beaconState.ShardCommitteesAtSlots,
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		// We determine the assigned shard ID for the validator
-		// based on a public key and current crystallized state.
-		shardID, err := v.ValidatorShardID(
-			val.PublicKey,
-			beaconState.ValidatorRegistry,
-			beaconState.ShardCommitteesAtSlots,
-		)
-		if err != nil {
-			return nil, err
-		}
-		assignments = append(assignments, &pb.Assignment{
-			PublicKey:    val,
-			ShardId:      shardID,
-			Role:         role,
-			AssignedSlot: assignedSlot,
-		})
-	}
-	return assignments, nil
-}
